# docs/api/openapi.yaml
openapi: '3.0.3'
info:
  title: Kalki-Mirror API
  version: 0.2.1
  description: >-
    REST API for the KALKI Tantrik Intelligence Platform.
    Pattern recognition, transit geometry, AI-powered Vedic content,
    and membership management.
  contact:
    name: KALKI Platform
    url: https://www.astrokalki.com
  license:
    name: Proprietary
servers:
  - url: https://www.astrokalki.com
    description: Production
...
paths:
...
  /api/ai/archetype-quiz:
    post:
      operationId: aiArchetypeQuiz
      tags: [AI]
      summary: AI archetype quiz analysis
      description: >-
        Analyzes archetype quiz answers to determine dominant patterns.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items:
                    type: string
                    minLength: 5
                  minItems: 5
                  maxItems: 10
      responses:
        '200':
          description: Archetype analysis result
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT via Bearer token (cookie-based).
        Used for standard authenticated routes.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT with ADMIN, SUPERADMIN, EDITOR, or REVIEWER role required.

  # ── Reusable Responses ──────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: LLM not configured or returned invalid JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Health
...